Ordersection

+++ src/main/java/nana/shop/online/model/SellerReport.java
@@ -32,12 +32,12 @@ public class SellerReport {
     @OneToOne
     private Seller seller;
     
-    private Long totalEarnings = 0L;
-    private Long totalSales = 0L;
-    private Long totalRefunds = 0L;
-    private Long totalTax = 0L;
-    private Long netEarnings = 0L;
-    
+    private double totalEarnings = 0L;
+    private double totalSales = 0L;
+    private double totalRefunds = 0L;
+    private double totalTax = 0L;
+    private double netEarnings = 0L;
+   
     private Integer totalOrders = 0;
     private Integer canceledOrders = 0;
     private Integer totalTransactions = 0;
